raise exception instead of returning 'error'
