+++ corehq/ex-submodules/couchforms/exceptions.py
@@ -1,3 +1,4 @@
+from couchforms.const import MAGIC_PROPERTY
 
 
 class CouchFormException(Exception):
@@ -17,3 +18,30 @@ class MissingXMLNSError(CouchFormException):
 
 class UnexpectedDeletedXForm(Exception):
     pass
+
+
+class BadRequest(Exception):
+    def __init__(self, message):
+        self.message = message
+
+
+class MultipartFilenameError(BadRequest):
+    def __init__(self):
+        super().__init__(
+            'If you use multipart/form-data, please name your file %s.\n'
+            'You may also do a normal (non-multipart) post '
+            'with the xml submission as the request body instead.\n' % MAGIC_PROPERTY
+        )
+
+
+class MultipartEmptyPayload(BadRequest):
+    def __init__(self):
+        super().__init__(
+            'If you use multipart/form-data, the file %s'
+            'must not have an empty payload\n' % MAGIC_PROPERTY
+        )
+
+
+class EmptyPayload(BadRequest):
+    def __init__(self):
+        super().__init__('Post may not have an empty body\n')
